[enum-simplification] Teach enum simplification how to hoist releases into switch regions.
This is done by checking if all predecessors of a BB have an enum case
associated with it in the enum simplification dataflow. If so, we move
the release value into the predecessors and put them onto the payload of
the enum (if the enum has one) or eliminate the enum release_value
completely if the enum case has no payload.

Thus the algorithm is actually more general than just the switch region
problem. But that is the main target, so it makes sense to keep it in
the main log title.

<rdar://problem/17139960>

Swift SVN r18774