SIL tools always operate as -parse-stdlib, even when working with modules.
r18796 added a check that the standard library is always available when
not working in -parse-stdlib or SIL mode, but sil-opt and sil-extract have
no need to load the stdlib if the module file they're working on does not
depend on it. Treat these tools as if they are always in a -parse-stdlib mode.

This fixes the failure in witness_tables.sil, which is due to us pulling in
witness tables from all loaded modules.

Swift SVN r18807