buildbot-script.sh: Factor out common settings, so that ASan can override them.
Our ASan compiler currently produces code that is not ABI-compatible with
the default compiler. This is probably a bug, but for now we can work around
it by building everything with the same compiler. This is a nice cleanup of
common flags as well.

Patch by Kuba Brecka!

Swift SVN r18838