Add an early binding pass for accelerating virtual method dispatch.
Dynamic languages are able to implement inline caches for virtual calls, but swift is statically compiled, so we have to guess the types at compile time. The early binding pass guesses that types at the bottom of the class hierarchy are not subclassed and emits direct calls to these passes. It converts class_method calls into the following code:

if (Instance is of time Foo) {
	Foo::ping()
} else {
	Instance->ping();
}

The check if an instance is of a specific type is inexpensive, it is simply a load+icmp sequence. 

Swift SVN r18860