[codemotion] Teach SILCodeMotion how to use the ARC infrastructure to be more aggressive in sinking retain insts over switch_enums.
This commit reuses utility methods from the ARC optimizer to know if an
instruction can decrement the ref count of a retain_value's operand. If
we can conservatively ascertain that it can not, we can ignore the
instruction for the purposes of determining if we can move the
retain_value into the switch region.

<rdar://problem/17225910>

Swift SVN r18894