Ensure that we don't emit debug info for function arguments twice.
There is an undocumented invariant that function arguments can only
described by a either a dbg.value or a dbg.declare. In the Dwarf backend
function arguments are inserted into an array that is indexed by ArgNo and
if more than one description exist it will overwrite the previous one, and
the previous DbgVariable will become an orphan as detected by the
assertion.

<rdar://problem/17315406>

Swift SVN r18903