Remove SWIFT_ENABLE_GLOBAL_ARC_OPTS option and add SWIFT_STDLIB_USE_ASSERT_CONFIG_RELEASE.
We no longer need the former. The latter is important since normally the
stdlib does not build with a release assertion configuration implying
that instead of being simple traps that are typically what the ARC
optimizer would pattern match, traps emit a message.

When dumping the stdlib, to see the true effect of the global arc
optimizer, one needs to turn on SWIFT_STDLIB_USE_ASSERT_CONFIG_RELEASE
since otherwise the ARC optimizer will not handle traps.

This is an ok restriction because when compiling non-stdlib programs
with optimization, the stdlib will be imported into a Release
environment (if not an -Ofast environment).

Swift SVN r18978