Address rdar://problem/17125445 ("dead _ in top level code crashes silgen")
DiscardAssignment expressions are special in that during constraint generation they'll introduce a new type variable, but not place any constraints upon it.  (They are the only expression kind that behaves in this way.) If no subsequent expressions constrain the type variable, we may end up with a failed constraint system that's devoid of constraints, and hence no information to synthesize a diagnostic from. With no diagnostic associated with the DiscardAssignmentExpr's source location, we'll attempt to generate SIL and raise an assertion failure. Fortunately, we can detect these cases during the constraint salvage phase, and raise an appropriate error.

Swift SVN r19020