Rework handling of optionals for bridging forced and conditional downcasts.
Previously, we were unable to handle bridged downcasts to optional
types from optional sources, because because we applied the bridging
operation after we had already evaluated all of the bound optionals
(causing a crash). Now, we perform the bridging immediately after the
underlying forced or conditional cast, before evaluating the outer
bound optionals.

This also eliminates a bunch of code duplication between the forced
and conditional downcasts, now that the bridging code is shared.



Swift SVN r19065