Fixed an assertion that would fire when debug info was made for an AssignExpr where the left hand side was implicit with an invalid source location.
The fix is to detect this case and return the source's full source location. Otherwise it returns the source location using the destination's start and source's end.

Also modified LLDB's swift build to build the tests for any builds with asserts enabled so we can run the swift tests to ensure no breakage in the future.



Swift SVN r19136