Devirtualizer: work around partially implemented feature in the type system
We fail devirtualizing a specialized protocol method because the number (1) of
substitutions obtained from looking in the witness table does not match up with
the number (2) of substitutions expected by the function signature substitution
code.

As a workaround until the type code is fixed, check for this condition and bail.

rdar://17399536

Swift SVN r19137