[CodeMotion] sink retains down so we can pair up retains with releases.
This enables us to pair up retains with releases on the fast path with inline
cache. The run time of rdar://17327077 is down to 0.07s from 0.7s.

We should be able to simplify the code for sinking retains down switch_enum,
now we have the enum simplification.

Thanks Michael for the review.


Swift SVN r19283