Add another pass of CFG simplification after inlining.
Inlining exposes more opportunities for CFG simplifications, and this
could be beneficial before ARC opts.

Because we create inline "caches" fairly late we also need this in order
to clean up redundant checked_cast_br instructions that are exposed as a
result of inlining since we only run the SSA passes once after the
inline cache pass.

The change to actually optimize the checked_cast_br is forthcoming.

Swift SVN r19557