[utils] add pre-commit-benchmark script
This script runs the benchmarks in benchmark/PrecommitBench/ against
your working tree and compares the result with the results of running
the benchmark against origin/master, or any commit you specify on the
command line.  Its output is roughly in CSV format.  It keeps a cache of
benchmark results, tagged with the working tree state, so chances are
good that it won't have to change your working copy state.  However, if
results for the specified commit are not in the cache, the script will
temporarily switch your working copy to that commit in order to build.

Timing parsing, comparison, and printing code by Andy Trick.  Thanks,
Andy!

Swift SVN r19646