Remove bogus assert and tighten up induction variable determination.
There was a (commented out) assert that the entry node to each SCC we
find would be a SILArgument. This is not always the case, so I removed
the assert and added a test case that demonstrates this.

I tightened up the IV recognition code a little bit as well.

Swift SVN r19885