[enum-simplification] Use .find not [] when accessing various maps.
This eliminates unneeded object allocations and allows me to just get rid of the
[Key] operator on PreallocatedMap. The reason why I want to get rid of that
operator is that these maps are meant for dataflow and it is very easy to
forget to check for unreachable predecessors.

The places I updated here can not have that problem, but I think the symmetry is
important for people who are unfamiliar with the file (i.e. use .find, not [])
in this file.

Swift SVN r19907