Skeleton for a SideCar data library.
This is completely non-functional at the moment, but will eventually
turn into a side-car data file that we can use to describe various
properties of C/Objective-C entities that can be fed into the Clang
importer. It's provided as a separate library because this information
might eventually be useful for Clang as well.

Swift SVN r19987